+++ assets/javascripts/kitten/components/layout/dashboard-layout/test.js
@@ -67,9 +67,7 @@ describe('<DashboardLayout />', () => {
             <DashboardLayout.Alerts>Alert!</DashboardLayout.Alerts>
 
             <p>Main content</p>
-            <DashboardLayout.Toaster>
-              Toaster
-            </DashboardLayout.Toaster>
+            <DashboardLayout.Toaster>Toaster</DashboardLayout.Toaster>
           </DashboardLayout>,
         )
         .toJSON()